Refactor FriendsViewControllerVM to improve architecture and functionality
- Fix search functionality by making searchText @Published for reactive UI updates
- Add filteredInviteFriends computed property for invitation list search support
- Extract common state management logic into private helper methods:
  * resetState() for complete state reset
  * resetLoadingState() for loading state reset
  * setSuccessState() for success state handling
  * setErrorState() for unified error handling
- Improve reloadFriendList() to properly handle invitation list reloading based on current scenario
- Move NetworkInterfaceWrapper to dedicated Utilities/NetworkInterfaceWrapper.swift file
- Eliminate code duplication and improve maintainability
- Ensure consistent error handling across all async methods
